Add files via upload
Create a class called "Person" that has the following instance variables:
firstName (String)
lastName (String)
age (int)
The class should have a constructor that takes in values for each of these instance variables and sets them accordingly. It should also have the following methods:
getFullName(): this method should return the full name of the person in the format "firstName lastName".
getAge(): this method should return the age of the person.
setAge(): this method should modify the age of the person.
canVote(): this method should return a boolean indicating whether or not the person is old enough to vote. In the United States, the voting age is 18.
Create two subclasses of "Person": "Student" and "Employee". The "Student" class should override the getFullName method to return the full name of the student in the format "lastName, firstName". The "Employee" class should override the getFullName method to return the full name of the employee in the format "firstName lastName (Employee)".
Create a class called "University" that that has the following instance variables:
persons (ArrayList of Person objects)
The class should have a constructor that takes in the value for this instance variable and sets it accordingly. It should also have the following methods:
getFullNames(): this method should return a comma separated String of the full names of each of the persons in the Arraylist.
getOldestPerson(): this method should return the oldest person. The return type of this method should be Person.
getYoungestPerson(): this method should return the youngest person. The return type of this method should be Person.
getVoters(): this method should return the persons who can vote. The return type of this method should be an Arraylist of Person.
getNumberOfEmployees(): this method should use the full name of the persons to count and return the number of employees in the Arraylist.
Finally, write a class named UniversityTest and add a main method to it to demonstrate all the methods of the University class.