Add files via upload
QUESTION----

Consider that you have to build a password manager application. The application takes the URL
of a website, username and password as input and stores it in an encrypted format. At first, the
user enters the application and chooses the New Login Info option from the console menu. Then,
he needs to enter the URL, username and password of the website. These will be stored in an
arraylist. The password must be stored in an encrypted format. In order to encrypt the password, it
is left shifted n times, where n is less than the password length. For example, if your password is
“hello” and n is 2, then the encrypted password is “llohe”. When the user wants to see the details
of the website, he will choose the View Login Info option. Then, he will enter the name of the
website. The application will show the username and password of the website in plain text
decrypted format. Note that, in order to decrypt the encrypted password, you can perform the
right shift n times. For example, if you right shift “llohe” 2 times, it will be “hello”.
Now, perform the following tasks.
a) Create a class named AuthenticationInfo which will contain the website URL, username and
password as fields and getter-setters for these fields. Create a constructor to initialize these
fields.

2

b) Create a class named EncryptionUtil that contains two methods: encrypt and decrypt. The
encrypt method receives a String and returns an encrypted version of that String. The
decrypt methods work in a similar way, except that it implements the decryption function.
2

c) Create a class named Application which has an ArrayList of AuthenticationInfo objects. It
also has a method named start that shows the menu in the console. When the New Login
Info option is selected and URL, username and password input is complete, it calls another
method named createLoginRecord which is also inside the Application class. The
createLoginRecord method receives an AuthenticationInfo object as parameter and returns
an AuthenticationInfo object by encrypting the password in that object. This Returned
AuthenticationInfo is stored inside the Arraylist. When the View Login Info option is
selected and the URL is inputted, a method in the Application class named viewLoginRecord
is called which receives a String URL and returns an AuthenticationInfo object where the
password is in a decrypted format.