Add files via upload
Person Search Project ( *KHOJ The SEARCH*)
This project is a simple Java program that demonstrates a person search functionality based on specific conditions. It allows you to search for persons based on their age and city of residence.

Features
The Person class represents a person with attributes such as firstname, lastname, age, phone number, and city.
The ISearcher interface defines a common search behavior with the Cheackcondition method.
The AgeSearcher class implements the ISearcher interface and searches for persons with a specific age condition.
The CitySeracher class implements the ISearcher interface and searches for persons living in a specific city.
The Main class contains the main method that demonstrates the usage of searchers and performs the person search.
Usage
Create instances of the Person class with different attributes.
Add the persons to an ArrayList.
Instantiate searchers (CitySeracher, AgeSearcher, etc.) based on the search criteria you want to apply.
Iterate over the persons in the ArrayList and use the searchers to filter and display the desired results.
Feel free to modify the search conditions, add additional searchers, or expand the functionality as needed.